Add files via upload
This poster highlights the Industrial Talk organized by Dr. Muhammad Iqbal Tariq, where representatives from Clarity Techworks were invited to discuss ICT jobs. I can say that the meeting was very interesting but the point that resonated with me was that jobs don't solely look for people with high GPAs but rather they look for individuals who have created a wide range of projects and that are proven to have the ability to solve problems.